tutor-agent/pyproject.toml
[project]
name = "tutor-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "firecrawl-py>=4.21.0",
    "grandalf==0.8",
    "langchain[openai]>=1.2.13",
    "langgraph>=1.1.3",
    "langgraph-checkpoint-sqlite==3.0.3",
    "langgraph-cli[inmem]>=0.4.19",
    "langgraph-supervisor>=0.0.31",
    "langgraph-swarm>=0.1.0",
    "pytest>=9.0.2",
    "python-dotenv==1.2.2",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]
